etl-script/requirements.txt
pandas==1.5.3
numpy==1.23.5
sqlalchemy==1.4.9
minio==7.1.14
python-dotenv==1.0.0
pyarrow==6.0.1
pymysql==1.1.0
